requirements.txt
crewai>=0.19.0
python-dotenv>=1.0.0 